requirements.txt
django
gunicorn
django-heroku
beautifulsoup4==4.8.1
certifi==2020.12.5
chardet==4.0.0
Django==2.2.7
django-ajax-search==1.5.1
django-bootstrap-modal-forms==1.5.0
django-bootstrap4==1.0.1
django-bower==5.2.0
django-channels==0.7.0
django-crispy-forms==1.8.0
django-nvd3==0.9.7
django-suit==0.2.28
django-widget-tweaks==1.4.8
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.2.0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2019.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==1.9.5
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.26.2
